requirements.txt
aiodns==3.0.0
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
black==22.3.0
cffi==1.15.0
click==8.1.3
Django==4.0.5
dnspython==1.15.0
dnspython3==1.15.0
idna==3.3
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycares==4.1.2
pycparser==2.21
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.18.3
sniffio==1.2.0
sqlparse==0.4.2
tomli==2.0.1
